+++ frontend/src/app/account/account-actions.tsx
@@ -3,7 +3,7 @@
 import { cookies } from 'next/headers';
 import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
-import { fetchWrapper } from '@/app/_lib/api';
+import { fetchWrapper } from '@/app/_lib/fetch-wrapper';
 
 async function processAuthenticationForm({formData, path}: {formData: FormData, path: string}) {
   const response = await fetchWrapper(
